Fix IFACE_CMDBUF struct misalignment on 64-bit systems causing session hang

+++ source/ifacedef.h
@@ -6,6 +6,8 @@
    updated: Mark Wahl DL4YBG 96/09/23
 */
 
+#include <stdint.h>
+
 #define HEAD_LEN sizeof(IFACE_HEADER)
 #define IFACE_PACLEN 256
 #define MAX_LEN (HEAD_LEN + IFACE_PACLEN)
@@ -31,7 +33,7 @@ typedef struct iface_header{
 #define LEN_BULLID (LEN_SIMPLE + sizeof(int) + BULLIDLEN + 1)
 #define LEN_SETRWMODE (LEN_SIMPLE + sizeof(int))
 #define LEN_BOXISBUSY (LEN_SIMPLE + sizeof(int))
-#define LEN_BCCALLBACK (LEN_SIMPLE + sizeof(long))
+#define LEN_BCCALLBACK (LEN_SIMPLE + sizeof(int32_t))
 #define LEN_SETUNPROTO (LEN_SIMPLE + 20)
 #define LEN_CONNECT (LEN_SIMPLE + sizeof(int) + 20)
 #define LEN_TNTRESPONSE (LEN_SIMPLE + sizeof(int))
@@ -76,7 +78,7 @@ typedef struct iface_cmdbuf{
     /* CMD_BOXISBUSY */
     int boxisbusy;
     /* CMD_BCCALLBACK */
-    long file_id; 
+    int32_t file_id; 
     /* CMD_SETUNPROTO */
     struct {
       char qrg[20];
@@ -166,7 +168,7 @@ typedef struct bcast_headinfo {
   char tnc;
   char port;
   char qrg[20];
-  long file_id;
+  int32_t file_id;
   unsigned short file_type;
   char filename[256];
   char address[256];
